+++ source/CHANGELOG.md
@@ -3,10 +3,13 @@
 
 Development
 -------------
+NEW:
+* Added an error text in the Time Series Properties window when a time series' volume is NaN 
+
 FIXED:
 * Fixed superfluous metadata in export to ZRXP format #126
 * Added a check for title uniqueness when a title is edited in the Time Series properties window
-* Disable the ability to sort columns in the Time Series Values window
+* Disabled the ability to sort columns in the Time Series Values window
 
 Version 2.6.0
 -------------